fix: satisfy workspace font-size budgets and localization checks
