+++ Sources/AppDelegate.swift
@@ -16221,8 +16221,6 @@ final class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCent
         return false
     }
 
-    // Shared by keyboard shortcuts and the socket command. Admission may queue
-    // behind a panel transfer; callers must not interpret it as completion.
     func enqueueWorkspaceTerminalFontSizeChange(
         _ action: KeyboardShortcutSettings.Action,
         workspace: Workspace,
